changed the hash value from uint to ulong to make it build on AIX
[git-p4: depot-paths = "//depot/qt/main/": change = 119447]

+++ src/tools/qhash.cpp
@@ -10,7 +10,7 @@
     "a", "aa", "aaa", "aaaa", ...
 */
 
-uint qHash(const QByteArray& key)
+ulong qHash(const QByteArray& key)
 {
     const uchar *p = (const uchar *) key.data();
     int n = key.size();
@@ -26,7 +26,7 @@ uint qHash(const QByteArray& key)
     return h;
 }
 
-uint qHash(const QString& key)
+ulong qHash(const QString& key)
 {
     const QChar *p = key.unicode();
     int n = key.length();
